+++ packages/tdee-explorer/gatsby-config.js
@@ -8,5 +8,15 @@ module.exports = {
     "gatsby-plugin-typescript",
     "gatsby-source-tdee-json-api",
     "gatsby-source-bin-day-api",
+    {
+      resolve: `gatsby-plugin-prefetch-google-fonts`,
+      options: {
+        fonts: [
+          {
+            family: `Actor`,
+          },
+        ],
+      },
+    },
   ],
 };
