+++ packages/tdee-explorer/package.json
@@ -13,11 +13,12 @@
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing \""
   },
   "dependencies": {
-    "@tdee/gsheet-log-fetcher": "0.0.2",
     "@tdee/bin-day-fetcher": "0.0.1",
+    "@tdee/gsheet-log-fetcher": "0.0.2",
     "@tdee/types": "0.0.1",
     "d3": "^5.12.0",
     "gatsby": "^2.17.10",
+    "gatsby-plugin-prefetch-google-fonts": "^1.4.3",
     "gatsby-plugin-sass": "^2.1.20",
     "gatsby-plugin-typescript": "^2.1.15",
     "node-sass": "^4.13.0",
