+++ packages/tdee-explorer/src/layouts/centered.module.css
@@ -1,0 +1,7 @@
+.centered {
+  box-sizing: border-box;
+  max-width: var(--measure);
+  margin-left: auto;
+  margin-right: auto;
+  align-items: center;
+}
