--- packages/tdee-explorer/src/pages/Index.module.scss
@@ -1,13 +1,0 @@
-.Container {
-  max-width: 60rem;
-  margin: 3rem auto 0;
-  font-family: sans-serif;
-
-  h1 {
-    color: #333;
-  }
-
-  p {
-    color: #aaa;
-  }
-}
