+++ packages/tdee-explorer/src/styles/global.css
@@ -1,0 +1,29 @@
+:root {
+  --ratio: 1.5;
+  --s0: 1rem;
+  --s1: calc(var(--s0) * var(--ratio));
+  --s2: calc(var(--s1) * var(--ratio));
+  --s3: calc(var(--s2) * var(--ratio));
+  --s4: calc(var(--s3) * var(--ratio));
+  --s5: calc(var(--s4) * var(--ratio));
+
+  --measure: 60ch;
+}
+
+* {
+  background-color: #f7f5e8;
+  box-sizing: border-box;
+  color: #050704;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica,
+    Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+}
+
+.logo {
+  border-bottom: 4px #dfcc4c solid;
+  font-family: Actor;
+  font-size: 32px;
+  line-height: 1;
+  max-width: 138px;
+  overflow-wrap: break-word;
+  padding-bottom: 20px;
+}
